==> 01_Import.Rmd <==
##Importing the data

The data was posted in an Excel file on google drive ([test roll results](https://drive.google.com/file/d/1oFhszSd7A210-YbTaEfz89A1eM7QgK5U/view?usp=sharing)). A quick test with the googlesheets package convinced me that I was better off just downloading the excel file.  I saved it in a ./data/raw/ folder.

### File Format
The Excel file consisted of 6 tabs, one for each die being tested.  Data was organized in a very excel fashion, with data and calculations included on the same tabs, and the data listed in cells A5:... It took a little creativity to get the right cells, and there were just enough tabs to copy/paste rather than building a formula to do it this time.

###Import

```{r Import, cache = TRUE, echo = FALSE, message = FALSE, warning = FALSE}

# Load libraries ----------------------------------------------------------

library(tidyverse)
#library(googlesheets)
library(readxl)

# read_name_file <- params$read_name_file

# Import the raw data in csv ----

read_path <- "./data/raw/"
write_path <- "./data/"




read_path_name_file <- paste0(read_path, read_name_file)

responses_num <- 0

#D20
die_sides = 20
die <- paste0("D", die_sides)
rolls <- read_xlsx(read_path_name_file,
          na = "",
          sheet = die,
          range = "D1:D2") %>% 
  .[[1]]
  
read_xlsx(read_path_name_file,
          na = "",
          sheet = die,
          range = paste0("A1:A", rolls + 5)) %>% 
  filter(as.integer(D20) %in% c(1:die_sides)) %>% 
  write_csv(paste0(write_path, die, ".csv")) %>%
  nrow() -> responses_num_d20


#D12
die_sides = 12
die <- paste0("D", die_sides)
rolls <- read_xlsx(read_path_name_file,
          na = "",
          sheet = die,
          range = "D1:D2") %>% 
  .[[1]]
  
read_xlsx(read_path_name_file,
          na = "",
          sheet = die,
          range = paste0("A1:A", rolls + 5)) %>% 
  filter(as.integer(D12) %in% c(1:die_sides)) %>% 
  write_csv(paste0(write_path, die, ".csv")) %>%
  nrow() -> responses_num_d12


#D10
die_sides = 10
die <- paste0("D", die_sides)
rolls <- read_xlsx(read_path_name_file,
          na = "",
          sheet = die,
          range = "D1:D2") %>% 
  .[[1]]
  
read_xlsx(read_path_name_file,
          na = "",
          sheet = die,
          range = paste0("A1:A", rolls + 5)) %>% 
  filter(as.integer(D10) %in% c(1:die_sides)) %>% 
  write_csv(paste0(write_path, die, ".csv")) %>%
  nrow() -> responses_num_d10


#D8
die_sides = 8
die <- paste0("D", die_sides)
rolls <- read_xlsx(read_path_name_file,
          na = "",
          sheet = die,
          range = "D1:D2") %>% 
  .[[1]]
  
read_xlsx(read_path_name_file,
          na = "",
          sheet = die,
          range = paste0("A1:A", rolls + 5)) %>% 
  filter(as.integer(D8) %in% c(1:die_sides)) %>% 
  write_csv(paste0(write_path, die, ".csv")) %>%
  nrow() -> responses_num_d8


#D6
die_sides = 6
die <- paste0("D", die_sides)
rolls <- read_xlsx(read_path_name_file,
          na = "",
          sheet = die,
          range = "D1:D2") %>% 
  .[[1]]
  
read_xlsx(read_path_name_file,
          na = "",
          sheet = die,
          range = paste0("A1:A", rolls + 5)) %>% 
  filter(as.integer(D6) %in% c(1:die_sides)) %>% 
  write_csv(paste0(write_path, die, ".csv")) %>%
  nrow() -> responses_num_d6


#D4
die_sides = 4
die <- paste0("D", die_sides)
rolls <- read_xlsx(read_path_name_file,
          na = "",
          sheet = die,
          range = "D1:D2") %>% 
  .[[1]]
  
read_xlsx(read_path_name_file,
          na = "",
          sheet = die,
          range = paste0("A1:A", rolls + 5)) %>% 
  filter(as.integer(D4) %in% c(1:die_sides)) %>% 
  write_csv(paste0(write_path, die, ".csv")) %>%
  nrow() -> responses_num_d4

#print(paste0(responses_num, " responses recorded for analysis, saved as ", write_path_name_responses))

# responses <- ls()[9:14]
# 
# for (i in 1:6) {
#   test[i] <- paste0("* ", responses[i], ": **`r ", responses[i], "`** rolls recorded")
# }
# test

```

I imported each sheet separately and saved just the data into 6 separate .csv files.  I like to do this so that tidying begins on a standard and fixed set of data.


* d4: **`r responses_num_d4`** rolls recorded
* d6: **`r responses_num_d6`** rolls recorded   
* d8: **`r responses_num_d8`** rolls recorded
* d10: **`r responses_num_d10`** rolls recorded 
* d12: **`r responses_num_d12`** rolls recorded
* d20: **`r responses_num_d20`** rolls recorded

That's a total of **`r responses_num_d4 + responses_num_d6 + responses_num_d8 + responses_num_d10 + responses_num_d12 + responses_num_d20`** rolls recorded.



```{r Import_cleanup, cache = TRUE, echo = FALSE, message = FALSE, warning = FALSE}
# rm(list = c("name_invitations_day",
#             "name_invitations_month",
#             "name_invitations_year",
#             "name_responses_day",
#             "name_responses_month",
#             "name_responses_year",
#             "read_name_invitations",
#             "read_name_responses",
#             "read_path",
#             "read_path_name_invitations",
#             "read_path_name_responses",
#             "write_name_start",
#             "write_path",
#             "write_path_name_invitations",
#             "write_path_name_responses",
#             "invitations_num",
#             "responses_num"))

#print("01_Import")
import_warnings <- warnings()

```

Warnings: There were **`r length(import_warnings)`** warnings during the import.  They were:
`r import_warnings`
