==> 04_Analyse.Rmd <==
##Analyze the results of u/VaraNiN’s dice rolls
### Testing the dice for balance based on rolls and expected values



```{r combined_Load, dependson = "Transform", cache = TRUE, echo = FALSE, message = FALSE, warning = FALSE}

# Load libraries ----------------------------------------------------------

library(tidyverse)
library(broom)
library(tm)
library(RColorBrewer)
library(wordcloud)
library(gridExtra)
library(ggalluvial)
library(NPS)


# Load Values objects for assignment and filtering ------------------------

source("./Values.R")
source("./functions.R")


# Initialize an empty list to store the results ----
results <- list()


# set parameters & global values ----

output_folder <- "output"

my_fill_color <- brewer.pal(4, "Blues")[3] #for histograms, etc.

#for plots and graphs
my_theme <- theme(
  panel.background = element_blank(),
  axis.line = element_line(size = .5)
)



# Load the transformed data ----

report <- readRDS("data/combined_rolls.rds")

report_names_class <- tibble( "names" = names(report), "class" = map(report, class))

```




```{r convert_report_to_combined, dependson = "combined_Load", cache = TRUE, echo = FALSE, message = FALSE, warning = FALSE}

# Convert report  ------------------------------------------

report_info <- paste0( "results from Title (n = ", nrow(report), ")")
if (report_info != "") {report_info <- paste0( report_info, " ")}

name_token <- "combined"
if (name_token != "") {name_token <- paste0( name_token, "_")}


#special columns that aren't factors by should be included in charts/grapsh, etc.
# special_cols <- c( 1:2)

# report_names_class %>%
#   rownames_to_column() %>%
#   mutate(special_cols = rowname %in% special_cols,
#          col_num_name = paste0(rowname, " - ", names)
#          ) %>%
#   filter(special_cols == T) %>%
#   .$col_num_name


report_responses_num <- nrow(report)
# report_responses_num <- report %>% nrow()

```

**`r report_responses_num`** responses were loaded for analysis.



```{r combined_analysis, dependson = "combined_Load", cache = TRUE, echo = FALSE, message = FALSE, warning = FALSE}


chisq <- report %>% 
  count(die,rolls) %>% 
  nest(-die) %>%
  mutate(test = map(data, ~ chisq.test(.$n)), 
    tidied = map(test, tidy)) %>% 
  unnest(tidied, .drop = TRUE) %>% 
  mutate(chisq_result_0.05 = if_else(p.value < 0.05, "Significant Difference", "Ok")
         )

summary <- report %>%
  group_by(die) %>% 
  summarize(roll_num = n(),
            roll_mean = mean(rolls)) %>% 
  mutate(sides = as.integer(sub("D","",die)),
         expected_rolls_per_side = roll_num/sides,
         expected_mean = (sides + 1 ) /2) %>% 
  left_join(chisq) %>% 
  select(die, sides, roll_num, expected_rolls_per_side, roll_mean, expected_mean, chisq_p.value = p.value, chisq_result_0.05) %>% 
  arrange(sides)





```








```{r combined_Histograms, dependson = "combined_Load", cache = TRUE, echo = FALSE, message = FALSE, warning = FALSE}



report %>%
  left_join(summary) %>% 
  mutate(sides = as.integer(sub("D","",die))) %>% 
  arrange(sides) %>% 
  ggplot(mapping = aes(x = rolls)) +
  geom_bar(fill = my_fill_color) +
  geom_hline(aes(yintercept = expected_rolls_per_side)) +
  facet_wrap(~sides, nrow = 3) +
  labs(x = "Die face rolled, \n*horizontal line is the expected rolls per side",
       title = "Count of rolls per die side, organized by die"
       ) +
  my_theme



```




### Chi-Square analysis based on rolls and expected values

```{r combined_pairwise_Overall_Satisfaction, dependson = "combined_Load", cache = TRUE, echo = FALSE, message = FALSE, warning = FALSE}


# Pairwise by Overall_Satisfaction -----------------------------------------------------------
summary %>% 
  mutate(roll_mean = round(roll_mean, 1),
         chisq_p.value = round(chisq_p.value,2)
         ) %>% 
  knitr::kable(format = "html") %>% 
  kableExtra::kable_styling(bootstrap_options = "striped", full_width = F) %>% 
  kableExtra::scroll_box(width = "100%", height = "100%")

# plot_calc_pairs(df = report, x = Column, y3 = Overall_Satisfaction, ex_label = '')


```



##Conclusions: The dice look balanced

With all p-values > 0.05, the first pass analysis results in a colloquial conclusion* that the dice are balanced.  
* lots of exceptions, details, and better explanations.  

##Future thoughts: 
I'm less concerned with misunderstanding of the implications of the p-values in this analysis, than I am with the conservative nature of the Chi-Square test for this kind of analysis.  The variable representing the dice sides can move along the variable continuum from Nominal to Ratio based on use in various games.  In the case of Dungeons and Dragons, the scale on the D20 is typically ordinal or interval from 2-19, with special exceptions for 1 and 20.  A dice that is imbalanced with respect to 1 or 20 in Dungeons and Dragons would be of particular concern.  The D12 graph above shows a potential example, with quite a few more 20s rolled than 1s.  But that's a conversation and analysis for another time.



## Continuing exercise:
Trying to model the power of the test.

```{r trials, dependson = "combined_Load", cache = TRUE, echo = FALSE, message = FALSE, warning = FALSE}
bias = 0. #percentage bias

summary <- summary %>%
  mutate(fair_side_perc = 1 / sides)
         
trials <- tribble()

for (k in 0:10) {
  
  bias = k / 10
  
  for (j in 1:100) {
    for (i in 1:nrow(summary)) {
      trials <- sample(x = 1:summary$sides[i], 
                       size = summary$roll_num[i], 
                       replace = T, 
                       prob = c(summary$fair_side_perc[i] * (1 + bias), 
                                rep(summary$fair_side_perc[i], 
                                    summary$sides[i] -2),
                                summary$fair_side_perc[i] * (1 - bias))
      ) %>%
        as_tibble(.) %>% 
        add_row(value = summary$sides[i]) %>% 
        rename(rolls = value) %>% 
        mutate(bias = bias,
               trial = j,
               die = paste0("D", summary$sides[i])) %>% 
        bind_rows(trials)
    }        
  }
}

```


## Another reddit exercise:

[link](https://www.reddit.com/r/Rlanguage/comments/a8bc9v/please_help_me_to_create_a_function_of_8_sided/)

Please help me to create a function of 8 sided bias die ( the question is in below)
Here is the question:

Roll a 8-sided biased die repeatedly. Define the ‘bias’ by yourself. Starting with roll k =1, let Nk denote the result of roll k. If Nk > k, then you will roll again; otherwise you stop. Let R denote the number of rolls.
Write a function: roll8( )
input: n (trials)
output: n outcomes and R value

```{r roll8, dependson = "combined_Load", cache = TRUE, echo = FALSE, message = FALSE, warning = FALSE}

roll8 <- function() {
  Nk = 8
  k = 1
  R = 0
  n = 0.
  
  bias = 0.125 #percentage bias
  prob <- c(1/8 * (1 + bias), 
            rep(1/8, #percentatage for fair sided die
                8 -2), #number of fair sides in this analysis
            1/8 * (1 - bias))
  
  
  
  while (Nk > k) {
    Nk <- sample(x = 1:8, 
                 size = 1, 
                 prob = prob
    )
    R <- R + 1
    n[R] <- Nk
  }
  
  print(paste0("Rolls: ", R))
  print("Results: ")
  print(n)
}


roll8()

roll8v2 <- function(trials) {
  
  n = 0 #create trial number vector
  
  bias = 0.125 #percentage bias
  prob <- c(1/8 * (1 + bias), 
            rep(1/8, #percentatage for fair sided die
                8 -2), #number of fair sides in this analysis
            1/8 * (1 - bias))
  
  
  for (trial in 1:trials ) {
    k = 1
    R = 0
    Nk = 8
    
    while (Nk > k) {
      Nk <- sample(x = 1:8, 
                   size = 1, 
                   prob = prob
      )
      R <- R + 1
    }
    
    print(paste0("trial ", trial, " rolls ", R))

    n[trial] <- R

  }

  print("Results: ")
  print(n)
}


roll8v2(3)



```
